requirements/requirements_dev.txt
alive_progress==3.1.4
argcomplete==3.1.1
catalyst==22.4
certifi>=2022.12.7 # not directly required, pinned by Snyk to avoid a vulnerability.
fastapi>=0.101.0
fiona>=1.9.1
flake8==6.1.0
geopy==2.4.0
imageio==2.31.3
inputimeout==1.0.4
internet-sabotage3==0.1.6
ipykernel==6.25.1
lightly==1.4.6
matplotlib==3.7.2
mlflow==2.6.0
mypy==1.5.1
myst_parser==2.0.0
nptyping==2.5.0
numba>=0.57.0 # not directly required but pinned to ensure Python 3.11 compatibility.
numpy==1.25.2
onnx==1.14.0
onnx2torch==1.5.11
overload==1.1
pandas==2.0.2
pre-commit==3.4.0
protobuf>=3.19.5 # not directly required, pinned by Snyk to avoid a vulnerability.
psutil==5.9.5
pygments>=2.7.4 # not directly required, pinned by Snyk to avoid a vulnerability.
pytest==7.4.0
pytest-cov==4.1.0
pytest-lazy-fixture==0.6.3
pyyaml==6.0.1
rasterio>=1.3.6
requests==2.31.0
scikit-learn==1.3.0
segmentation-models-pytorch==0.3.3
setuptools==67.8.0
sphinx==6.2.1
sphinx-rtd-theme==1.3.0
starlette>=0.25.0 # not directly required, pinned by Dependabot to avoid a vulnerability.
tabulate==0.9.0
tensorflow==2.13.0
timm==0.9.2
torchgeo@git+https://github.com/microsoft/torchgeo
torchinfo==1.8.0
tox==4.8.0
tqdm==4.66.1
types-PyYAML==6.0.12.11
types-requests==2.31.0.2
types-tabulate==0.9.0.3
wandb==0.15.8
Werkzeug>=2.2.3 # Patches a potential security vulnerability.
wheel>=0.38.0 # not directly required, pinned by Snyk to avoid a vulnerability.
